feat: simple framework setup
